Added error handling if the table doesn't exist.
# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ dbConnection.py
@@ -4,7 +4,7 @@
 class Database:
 	import sqlite3 as lite
 	
-	#TODO Should be able to pass a connection object to this class. 
+	#TODO Maybr should be able to pass a connection object to this class. 
 	
 	def __init__(self, database):
 		if database: self.database = database
@@ -41,16 +41,16 @@ class Database:
 		if self.con:
 			with self.con:
 				cur = self.con.cursor()	
-				
-				query = "CREATE TABLE %s("%(tableName)
-				for column, dataType in listOfColsAndTypes.items():
-					if not query.endswith('('): query += ','	# append a comma to seperat values
-					query += "%s %s"%(column, dataType)
-				query += ")"
-				
-				if self.debug:print query 
-				
-				cur.execute(query)
+				if self.tableExists(cur, tableName):
+					query = "CREATE TABLE %s("%(tableName)
+					for column, dataType in listOfColsAndTypes.items():
+						if not query.endswith('('): query += ','	# append a comma to seperat values
+						query += "%s %s"%(column, dataType)
+					query += ")"
+					
+					if self.debug:print query 
+					
+					cur.execute(query)
 		else:
 			noConnection()
 			
@@ -58,17 +58,19 @@ class Database:
 		if self.con:
 			with self.con:
 				cur = self.con.cursor()
-				cur.execute("DROP TABLE %s"%(tableName))
-				if self.debug:print "Table %s was dropped."%(tableName)
-				
+				if self.tableExists(cur, tableName):
+					cur.execute("DROP TABLE %s"%(tableName))
+					if self.debug:print "Table %s was dropped."%(tableName)
+					
 	def selectFrom(self, tableName): 
 		#TODO look into row factory and text factory 
 		if self.con:
 			with self.con:
 				cur = self.con.cursor()
-				for row in cur.execute("SELECT * FROM %s"%(tableName)):
-					if self.debug:print row
-								
+				if self.tableExists(cur, tableName):
+					for row in cur.execute("SELECT * FROM %s"%(tableName)):
+						if self.debug:print row
+									
 	def insertInto(self, tableName, values):
 		# TODO allow insert of one or many columns in any order..
 		if self.con:
@@ -76,32 +78,37 @@ class Database:
 				cur = self.con.cursor()
 				#cur.execute("SELECT 1 FROM sqlite_master WHERE type='table' AND name='%s'"%(tableName))
 				#if cur.fetchone() == 1:
-				
-				query = "INSERT INTO %s ("%(tableName)
-				data = ") VALUES ("				
-				for column, value in values.items():
-					if not query.endswith('('): 
-						query += ','	# append a comma to seperat values
-						data += ','
+				if self.tableExists(cur, tableName):
+					query = "INSERT INTO %s ("%(tableName)
+					data = ") VALUES ("				
+					for column, value in values.items():
+						if not query.endswith('('): 
+							query += ','	# append a comma to seperat values
+							data += ','
+							
+						query += "%s"%(column)					
+						if isinstance(value,str):data += "'%s'"%(value)
+						else:data += "%s"%(value)
 						
-					query += "%s"%(column)					
-					if isinstance(value,str):data += "'%s'"%(value)
-					else:data += "%s"%(value)
+					query += data + ')'
 					
-				query += data + ')'
-				
-				if self.debug:print query
-				
-				cur.execute(query)
-				self.con.commit()
-	
+					if self.debug:print query
+					
+					cur.execute(query)
+					self.con.commit()
+		
 	def noConnection(self):
 		#Something should happen if there is no connection. 
 		if self.debug is 1:print 'There is no connection, attempting to connect.'
 		connectToDB(self)
 	
-	def tableExists(self,tableName):
-		cur.execute("SELECT 1 FROM sqlite_master WHERE type='table' AND name='%s'"%(tableName))
-		if cur.fetchone() == 1: return 1
-		return 0
+	def tableExists(self, cur, tableName):
+		try:			
+			cur.execute("SELECT 1 FROM sqlite_master WHERE type='table' AND name='%s'"%(tableName))
+			if cur.fetchone()[0] == 1: 
+				return True
+		except Exception, e: 
+			if self.debug:print "An Sql error occured: ",e.args[0]	
+			print "Table:",tableName," doesn't exist"
+		return False
